+++ CHANGES.md
@@ -6,6 +6,7 @@ v1.0.9 (TBD)
 Non-breaking changes:
 ---------------------
 * Process all the files in a directory as multiple input files through the new arguments `--input-directory` and `--output-directory`
+* New `--move-input-files` argument to move the input files to the output directory after processing
 * `--output` moved back as a shared argument (relevant for scripts using delimited2fixedwidth as an imported library), partially reverting a change introduced in v1.0.2
 * Ensure the `--overwrite-file` argument is only used with `--output`